requirements.txt
selenium
python-vlc
requests
tk
